Inline GeneratedGristCost
