+++ src/main/java/com/mraof/minestuck/alchemy/recipe/generator/ContainerGristCost.java
@@ -2,46 +2,55 @@ package com.mraof.minestuck.alchemy.recipe.generator;
 
 import com.google.gson.JsonObject;
 import com.mojang.serialization.JsonOps;
-import com.mraof.minestuck.alchemy.recipe.GeneratedGristCostCache;
+import com.mraof.minestuck.alchemy.recipe.GristCostRecipe;
+import com.mraof.minestuck.alchemy.recipe.SimpleGristCost;
 import com.mraof.minestuck.api.alchemy.GristSet;
+import com.mraof.minestuck.api.alchemy.GristType;
 import com.mraof.minestuck.api.alchemy.ImmutableGristSet;
 import com.mraof.minestuck.item.crafting.MSRecipeTypes;
+import com.mraof.minestuck.jei.JeiGristCost;
 import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.GsonHelper;
+import net.minecraft.world.Container;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.item.crafting.RecipeSerializer;
+import net.minecraft.world.level.Level;
 import net.minecraftforge.registries.ForgeRegistries;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
+import java.util.Collections;
+import java.util.List;
+import java.util.function.BiConsumer;
 
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
-public class ContainerGristCost extends GeneratedGristCost
+public final class ContainerGristCost extends SimpleGristCost
 {
 	private static final Logger LOGGER = LogManager.getLogger();
 	
-	private final ImmutableGristSet addedCost;
+	private final GeneratedGristCostCache cache;
 	
 	public ContainerGristCost(ResourceLocation id, Ingredient ingredient, ImmutableGristSet addedCost, @Nullable Integer priority)
 	{
 		super(id, ingredient, priority);
-		this.addedCost = addedCost;
+		this.cache = new GeneratedGristCostCache(context -> this.generateCost(context, addedCost));
 	}
 	
 	private ContainerGristCost(ResourceLocation id, Ingredient ingredient, @Nullable Integer priority, GeneratedGristCostCache cache)
 	{
-		super(id, ingredient, priority, cache);
-		this.addedCost = null;
+		super(id, ingredient, priority);
+		this.cache = cache;
 	}
 	
-	@Override
-	protected GristSet generateCost(GenerationContext context)
+	@Nullable
+	private GristSet generateCost(GenerationContext context, ImmutableGristSet addedCost)
 	{
 		ItemStack container = ingredient.getItems().length > 0 ? ingredient.getItems()[0].getCraftingRemainingItem() : ItemStack.EMPTY;
 		if(!container.isEmpty())
@@ -71,10 +80,36 @@ public class ContainerGristCost extends GeneratedGristCost
 		return MSRecipeTypes.CONTAINER_GRIST_COST.get();
 	}
 	
-	public static class Serializer extends GeneratedCostSerializer<ContainerGristCost>
+	@Override
+	public GristSet getGristCost(ItemStack input, @Nullable GristType wildcardType, boolean shouldRoundDown, @Nullable Level level)
+	{
+		return GristCostRecipe.scaleToCountAndDurability(cache.getCachedCost(), input, shouldRoundDown);
+	}
+	
+	@Override
+	public boolean matches(Container inv, Level level)
+	{
+		return cache.getCachedCost() != null && super.matches(inv, level);
+	}
+	
+	@Override
+	public void addCostProvider(BiConsumer<Item, GeneratedCostProvider> consumer)
+	{
+		GristCostRecipe.addCostProviderForIngredient(consumer, this.ingredient, this.cache);
+	}
+	
+	@Override
+	public List<JeiGristCost> getJeiCosts(Level level)
+	{
+		if(cache.getCachedCost() != null)
+			return Collections.singletonList(new JeiGristCost.Set(ingredient, cache.getCachedCost()));
+		else return Collections.emptyList();
+	}
+	
+	public static class Serializer extends AbstractSerializer<ContainerGristCost>
 	{
 		@Override
-		protected ContainerGristCost read(ResourceLocation recipeId, JsonObject json, Ingredient ingredient, Integer priority)
+		protected ContainerGristCost read(ResourceLocation recipeId, JsonObject json, Ingredient ingredient, @Nullable Integer priority)
 		{
 			ImmutableGristSet cost = ImmutableGristSet.MAP_CODEC.parse(JsonOps.INSTANCE, GsonHelper.getAsJsonObject(json, "grist_cost"))
 					.getOrThrow(false, LOGGER::error);
@@ -82,9 +117,18 @@ public class ContainerGristCost extends GeneratedGristCost
 		}
 		
 		@Override
-		protected ContainerGristCost create(ResourceLocation recipeId, FriendlyByteBuf buffer, Ingredient ingredient, int priority, GeneratedGristCostCache cache)
+		protected ContainerGristCost read(ResourceLocation recipeId, FriendlyByteBuf buffer, Ingredient ingredient, int priority)
 		{
+			GeneratedGristCostCache cache = GeneratedGristCostCache.read(buffer);
+			
 			return new ContainerGristCost(recipeId, ingredient, priority, cache);
+		}
+		
+		@Override
+		public void toNetwork(FriendlyByteBuf buffer, ContainerGristCost recipe)
+		{
+			super.toNetwork(buffer, recipe);
+			recipe.cache.write(buffer);
 		}
 	}
 }
